{{cookiecutter.project_dirname}}/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.ini
#
appdirs==1.4.3            # via black
asgiref==3.2.3            # via django
aspy.yaml==1.3.0          # via pre-commit
attrs==19.3.0             # via black, flake8-bugbear, flake8-mypy, pytest
backcall==0.1.0           # via ipython
black==19.10b0
cfgv==2.0.1               # via pre-commit
click==7.0                # via black, pip-tools
coverage==5.0
decorator==4.4.1          # via ipython, traitlets
dj-database-url==0.5.0
dj-email-url==0.2.0
django-configurations==2.2
django-debug-toolbar==2.1
django==3.0.1
entrypoints==0.3          # via flake8
filelock==3.0.12          # via tox
flake8-bugbear==19.8.0
flake8-docstrings==1.5.0
flake8-isort==2.8.0
flake8-mypy==17.8.0
flake8==3.7.9
identify==1.4.8           # via pre-commit
importlib-metadata==1.3.0  # via pluggy, pre-commit, pytest, tox
invoke==1.3.0
ipdb==0.12.3
ipython-genutils==0.2.0   # via traitlets
ipython==7.10.2           # via ipdb
isort[pyproject]==4.3.21
jedi==0.15.1
mccabe==0.6.1             # via flake8
more-itertools==8.0.2     # via pytest, zipp
mypy-extensions==0.4.3    # via mypy
mypy==0.760
nodeenv==1.3.3            # via pre-commit
packaging==19.2           # via pytest, tox
parso==0.5.2              # via jedi
pathspec==0.6.0           # via black
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==4.3.0
pluggy==0.13.1            # via pytest, tox
pre-commit==1.20.0
prompt-toolkit==3.0.2     # via ipython
psycopg2==2.8.4
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest, tox
pycodestyle==2.5.0        # via flake8
pydocstyle==5.0.1         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pygments==2.5.2           # via ipython
pyparsing==2.4.5          # via packaging
pytest-django==3.7.0
pytest==5.3.2             # via pytest-django
python-dotenv==0.10.3
pytz==2019.3              # via django
pyyaml==5.2               # via aspy.yaml, pre-commit
regex==2019.12.19         # via black
rope==0.14.0
six==1.13.0               # via cfgv, django-configurations, packaging, pip-tools, pre-commit, tox, traitlets
snowballstemmer==2.0.0    # via pydocstyle
sqlparse==0.3.0           # via django, django-debug-toolbar
testfixtures==6.10.3      # via flake8-isort
toml==0.10.0              # via black, isort, pre-commit, tox
tox==3.14.2
traitlets==4.3.3          # via ipython
typed-ast==1.4.0          # via black, mypy
typing-extensions==3.7.4.1  # via mypy
uwsgi==2.0.18
virtualenv==16.7.9        # via pre-commit, tox
wcwidth==0.1.7            # via prompt-toolkit, pytest
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
